Renamed coffee chat commands and edited their descriptions
