+++ src/commands/coffeeChat/coffeeChatTest.ts
@@ -2,16 +2,16 @@ import { Message, MessageEmbed } from 'discord.js';
 import { Args, Command, CommandOptions } from '@sapphire/framework';
 import { ApplyOptions } from '@sapphire/decorators';
 import { container } from '@sapphire/pieces';
-import { testPerformance } from '../../components/coffeechat';
+import { testPerformance } from '../../components/coffeeChat';
 import { EMBED_COLOUR } from '../../utils/embeds';
 
 @ApplyOptions<CommandOptions>({
-  aliases: ['coffee-test'],
-  description: 'Tests coffeematch.',
+  aliases: ['coffee-test', 'coffeetest', 'coffee-chat-test'],
+  description: 'Tests coffeechatmatch, displaying the number of rounds before there is a duplicate.',
   detailedDescription: `**Examples:**\n\`${container.client.options.defaultPrefix}coffeetest 10\``,
   requiredUserPermissions: ['ADMINISTRATOR']
 })
-export class CoffeeTestCommand extends Command {
+export class CoffeeChatTestCommand extends Command {
   async messageRun(message: Message, args: Args): Promise<Message> {
     // Mandatory argument is size
     const size = await args.rest('integer').catch(() => 'please enter a valid number of test users.');
